use /proc/self/statm to get current usage

+++ server/startup.go
@@ -79,7 +79,7 @@ func Serve(ctx *cli.Context) error {
 	// redirectsCache stores redirects in _redirects files
 	redirectsCache := cache.NewInMemoryCache()
 	// clientResponseCache stores responses from the Gitea server
-	clientResponseCache := cache.NewInMemoryCacheWithLimit(ctx.Uint64("memory-limit"))
+	clientResponseCache := cache.NewInMemoryCacheWithLimit(ctx.Int("memory-limit"))
 
 	giteaClient, err := gitea.NewClient(cfg.Gitea, clientResponseCache)
 	if err != nil {
